# Conflicts:
#	tkpdtransaction/src/main/java/com/tokopedia/transaction/purchase/adapter/TxVerAdapter.java
#	tkpdtransaction/src/main/java/com/tokopedia/transaction/purchase/fragment/TxListFragment.java

+++ tkpdcore/src/main/java/com/tokopedia/core/talkview/adapter/InboxTalkViewAdapter.java
@@ -178,17 +178,16 @@ public class InboxTalkViewAdapter extends TalkViewAdapter{
             @Override
             public boolean onMenuItemClick(MenuItem item) {
                 DialogFragment dialog;
-                switch (item.getItemId()) {
-                    case R2.id.action_delete:
-                        dialog = DeleteTalkDialog.createInstance(deleteListener(talk, position));
-                        dialog.show(fragment.getFragmentManager(), DeleteTalkDialog.FRAGMENT_TAG);
-                        return true;
-                    case R2.id.action_report:
-                        dialog = ReportTalkDialog.createInstance(reportListener(talk, position));
-                        dialog.show(fragment.getFragmentManager(), ReportTalkDialog.FRAGMENT_TAG);
-                        return true;
-                    default:
-                        return false;
+                if (item.getItemId() == R.id.action_delete) {
+                    dialog = DeleteTalkDialog.createInstance(deleteListener(talk, position));
+                    dialog.show(fragment.getFragmentManager(), DeleteTalkDialog.FRAGMENT_TAG);
+                    return true;
+                } else if (item.getItemId() == R.id.action_report) {
+                    dialog = ReportTalkDialog.createInstance(reportListener(talk, position));
+                    dialog.show(fragment.getFragmentManager(), ReportTalkDialog.FRAGMENT_TAG);
+                    return true;
+                } else {
+                    return false;
                 }
             }
